+++ server/model/Cannon.h
@@ -27,8 +27,8 @@ class Projectile : public Movable {
     public:
         Projectile(const std::string& name,
                    int damage,
-                   int velocity_x, int velocity_y,
-                   const std::vector<int>& initial_position);
+                   float velocity_x, float velocity_y,
+                   const std::vector<float>& initial_position);
         const std::string& get_name();
         void collide_with(Enemy* enemy);
         void collide_with(Object* object);
@@ -44,43 +44,43 @@ class Projectile : public Movable {
 
 class Plasma : public Projectile {
     public:
-        Plasma(int damage, int velocity_x, int velocity_y,
-               const std::vector<int>& initial_position);
+        Plasma(int damage, float velocity_x, float velocity_y,
+               const std::vector<float>& initial_position);
         void tick();
 };
 
 class Bomb : public Projectile {
     public:
-        Bomb(int damage, int velocity_x, int velocity_y,
-             const std::vector<int>& initial_position);
+        Bomb(int damage, float velocity_x, float velocity_y,
+             const std::vector<float>& initial_position);
         void tick();
 };
 
 class Magnet : public Projectile {
     public:
-        Magnet(int damage, int velocity_x, int velocity_y,
-               const std::vector<int>& initial_position);
+        Magnet(int damage, float velocity_x, float velocity_y,
+               const std::vector<float>& initial_position);
         void tick();
 };
 
 class Spark : public Projectile {
     public:
-        Spark(int damage, int velocity_x, int velocity_y,
-              const std::vector<int>& initial_position);
+        Spark(int damage, float velocity_x, float velocity_y,
+              const std::vector<float>& initial_position);
         void tick();
 };
 
 class Fire : public Projectile {
     public:
-        Fire(int damage, int velocity_x, int velocity_y,
-             const std::vector<int>& initial_position);
+        Fire(int damage, float velocity_x, float velocity_y,
+             const std::vector<float>& initial_position);
         void tick();
 };
 
 class Ring : public Projectile {
     public:
-        Ring(int damage, int velocity_x, int velocity_y,
-             const std::vector<int>& initial_position);
+        Ring(int damage, float velocity_x, float velocity_y,
+             const std::vector<float>& initial_position);
         void tick();
 };
 
@@ -92,7 +92,7 @@ class Ammo {
 
     public:
         Ammo(const std::string& name, int max);
-        Projectile* use(const std::vector<int>& position);
+        Projectile* use(const std::vector<float>& position);
         ~Ammo();
 };
 
@@ -105,7 +105,7 @@ class Cannon {
         Cannon();
         void receive_new_ammo(std::string& name);
         void change_current_ammo(unsigned int ammo_id);
-        Projectile* shoot(const std::vector<int>& position);
+        Projectile* shoot(const std::vector<float>& position);
         ~Cannon();
 };
