+++ server/model/Movable.h
@@ -8,37 +8,37 @@
 
 class Movable : public GameObject {
     protected:
-        const int velocity_x, velocity_y;
+        const float velocity_x, velocity_y;
 
     public:
-        Movable(const std::vector<int>& position,
-                const int velocity_x, const int velocity_y);
-        Movable(const int x, const int y,
-                const int velocity_x, const int velocity_y);
+        Movable(const std::vector<float>& position,
+                const float velocity_x, const float velocity_y);
+        Movable(const float x, const float y,
+                const float velocity_x, const float velocity_y);
         virtual void move();
-        virtual void correct_position(const std::vector<int>& obstacle_pos,
+        virtual void correct_position(const std::vector<float>& obstacle_pos,
                                       int obstacle_side);
         virtual ~Movable();
 };
 
 class UserMovable : public Movable {
     private:
-        const int gravity;
+        const float gravity;
         int direction_x, direction_y;
-        int current_vel_x, current_vel_y;
+        float current_vel_x, current_vel_y;
         bool on_stairs;
 
     public:
-        explicit UserMovable(const int velocity_x, const int velocity_y);
+        explicit UserMovable(const float velocity_x, const float velocity_y);
         void change_x_movement(bool pressed, bool forward);
         void change_y_movement(bool pressed, bool forward);
         void standing_on_stairs();
         void move();
         void reset_movement();
         void reset_position();
-        void correct_position(const std::vector<int>& obstacle_pos,
+        void correct_position(const std::vector<float>& obstacle_pos,
                               int obstacle_side);
-        std::vector<int> get_position();
+        std::vector<float> get_position();
         ~UserMovable();
 };
